+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 - Added support for service version 2026-04-06.
+- Added support for improved error handling for file share provisioning.
 
 ## 12.24.0b1 (Unreleased)